requirements.txt
pandas
numpy
langchain
langchain_openai
langchain-ollama
langchain_community
ollama
docx2txt
pypdf
mysql-connector-python
sqlalchemy
python-dotenv
scikit-learn
